+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2023-03-01
+
+### Added
+
+- IP service add /geolocate, /vpn, /domain and /proxy endpoints support
+
+## [1.3.0] - 2023-02-28
+
+### Added
+
+- Tenant ID support in Audit Service
+
 ## [1.2.0] - 2023-02-03
 
 ### Added
@@ -58,7 +70,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact client
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.2.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.4.0...main
+[1.4.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.3.0...v1.4.0
+[1.3.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/pangeacyber/pangea-go/compare/pangea-sdk/v1.1.0...v1.1.1